# docs/openapi/tradeguard-openapi.yaml
openapi: 3.0.3
info:
  title: TradeGuard 交易风控中枢 API 契约
  version: 1.0.0
  description: |
    统一 OpenAPI 3.0 契约，覆盖两类接口：
    1. web-api REST 接口（/api/**，供 web-portal 前端与演示调用，见 04 §10.1）；
    2. MCP 工具契约（AA-MCP-01/02 的 6+3 个工具，MCP 为 JSON-RPC 传输，
       其输入/输出 Schema 统一在 components.schemas 中以 x-mcp-tool 标注，
       与 REST 接口共用同一套组件，保证"所有接口统一 OpenAPI 3.0 规范"）。
    追溯：业务规则 BA-BR-x（01 §5）、场景 SC-x（06 §2）、数据字典 DD（08）。
servers:
  - url: http://localhost:8200
    description: web-api（docker compose 演示环境，宿主端口映射 8200→8000）
...
security:
  - bearerAuth: []
paths:
...
  /api/cases:
    get:
      tags: [事件工作台]
      operationId: listCases
      summary: 风险事件列表（真实读 DA-T-03）
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/CaseStatus' } }
        - { name: risk_min, in: query, schema: { type: integer, minimum: 0, maximum: 100 } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: 分页列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CaseSummary' }
  /api/cases/{case_id}:
    get:
      tags: [事件工作台]
      operationId: getCase
      summary: 事件详情（含共享状态 context，真实读 DA-T-03）
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/cases/{case_id}/signals:
    get:
      tags: [事件工作台]
      operationId: listSignals
      summary: 事件信号清单（真实读 DA-T-04）
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      responses:
        '200':
          description: 信号列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RiskSignal' }
  /api/cases/{case_id}/aggregate:
    post:
      tags: [事件工作台]
      operationId: aggregateCase
      summary: 触发信号聚合（API-W-17，AA-SK-01 确定性内核，SC-01/SC-11 载体）
      description: |
        立案→聚合（四源采集+防腐层翻译+降噪合并+velocity BA-BR-14+加权评分）→分级裁决：
        零信号降噪 ARCHIVED；低风险小额（风险分<40 且涉案<5000，BA-BR-01）自动放行 DISPOSED（BA-CAP-05）；
        其余转调查 INVESTIGATING；全源失败 E-AGG-ALL-FAIL 转人工。信号经 AA-MCP-01 落 DA-T-04（tg_app 写角色 DA-INV-05）。
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      responses:
        '200':
          description: 聚合完成
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AggregateResult' }
        '404':
          description: 事件不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: 状态冲突（非 REGISTERED/AGGREGATING 或全源失败转人工）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/cases/{case_id}/evidence:
    get:
      tags: [事件工作台]
      operationId: listEvidence
      summary: 证据链（真实读 DA-T-05，审批门户展示依据）
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      responses:
        '200':
          description: 证据列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Evidence' }
  /api/cases/{case_id}/dispositions:
    get:
      tags: [事件工作台]
      operationId: listDispositions
      summary: 处置执行记录（API-W-22，真实读 DA-T-06，按执行时序）
      description: 前端核验入口（API-W-19）依此取最近一条 executed 凭证作为 exec_id。
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      responses:
        '200':
          description: 处置记录列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Disposition' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/cases/{case_id}/graph:
    get:
      tags: [事件工作台]
      operationId: getRelatedGraph
      summary: 关联网络图谱（经 mcp-core query_related_graph，与 Agent 同源）
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
        - { name: hops, in: query, schema: { type: integer, default: 2, maximum: 2 } }
      responses:
        '200':
          description: 图谱
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GraphResponse' }
  /api/cases/{case_id}/review:
    post:
      tags: [事件工作台]
      operationId: submitReview
      summary: 中风险人工复核结论（SC-10 载体，BA-BR-01 中风险分段）
      description: |
        写 DA-T-03 状态 + 审计留痕（BA-BR-09），actor 取 X-Operator 头（human_only 守卫）。
        release→ReviewDismissed 排除欺诈归档 ARCHIVED；
        block/escalate→委托 AA-SK-03 review_confirm 自动建处置审批工单（API-M-11，US-E4-05），
        escalate 额外审计标记升级。
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewRequest' }
      responses:
        '200':
          description: 复核完成
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseSummary' }
        '409':
          description: 状态冲突（非 MANUAL_REVIEW 状态）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/cases/{case_id}/investigate:
    post:
      tags: [事件工作台]
      operationId: investigateCase
      summary: 触发欺诈调查（API-W-18，AA-SK-02 确定性内核，US-E4-01~03 载体）
      description: |
        假设匹配（规则兜底 + DA-KB-01 检索引用 doc_id，未命中显式声明）→
        fn_related_graph 2 跳扩展（BA-BR-06 命中黑名单 +30 幂等，API-M-13）→
        影响面统计 → 证据固化 DA-T-05（API-M-12）→ InvestigationCompleted 转 PENDING_APPROVAL。
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      responses:
        '200':
          description: 调查完成（假设/图谱/影响面/证据固化）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvestigationResult' }
        '404':
          description: 事件不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: 状态冲突（非 INVESTIGATING）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/cases/{case_id}/verify:
    post:
      tags: [事件工作台]
      operationId: verifyCase
      summary: 触发结果核验（API-W-19，AA-SK-04 确定性内核，US-E6-01/02 载体）
      description: |
        三分支裁决（B3）：
        1. 执行一致（无论 trace 是否完整）→VerificationPassed→VERIFIED→CaseArchived→ARCHIVED，
           审计报告落 DA-T-05，复盘摘要提入库申请（pending）；trace 缺口仅审计告警
           verification.trace_gap，绝不回滚一致执行；
        2. 不一致→VerificationFailed→反向处置（幂等键 :rollback 后缀，凭证按原动作+逆对
           兜底查询，C1 逆动作对授权）→RollbackExecuted→MANUAL_REVIEW + verification.p0 审计；
        3. 不一致且反向处置被拒/失败→RollbackEscalated→MANUAL_REVIEW +
           verification.escalated 审计（未执行反向不写 P0 已回滚）。
      parameters:
        - { $ref: '#/components/parameters/CaseIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyRequest' }
      responses:
        '200':
          description: 核验完成
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationResult' }
        '404':
          description: 事件或执行凭证不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: 状态冲突（非 DISPOSED）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/approvals:
    get:
      tags: [审批门户]
      operationId: listApprovals
      summary: 待审批/已审批队列（真实读 DA-T-07，BA-BR-13 超时标记）
      parameters:
        - { name: decision, in: query, schema: { type: string, enum: [pending, approved, rejected] } }
      responses:
        '200':
          description: 审批列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Approval' }
  /api/approvals/{approval_id}/decide:
    post:
      tags: [审批门户]
      operationId: decideApproval
      summary: 批准/驳回（02 §7 状态机人类触发入口，SC-02/SC-03 载体）
      description: |
        回填 DA-T-07 decision（回填后不可变更）→ 发布 ApprovalApproved/ApprovalRejected
        事件（actor=真实操作人）→ 审计留痕。驳回触发 BA-BR-07 回滚人工复核。
      parameters:
        - { name: approval_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecideRequest' }
      responses:
        '200':
          description: 决策完成
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '409':
          description: 工单已决策（幂等返回首次结果）或已过期
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/kb/applications/{doc_id}/publish:
    post:
      tags: [知识库后台]
      operationId: publishKbDocument
      summary: 确认发布（SC-05 载体，DA-INV-06 人工门控，仅策略管理员）
      parameters:
        - { $ref: '#/components/parameters/DocIdPath' }
      responses:
        '200':
          description: 已发布并触发向量化
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbDocument' }
        '409':
          description: 状态非 pending
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/kb/applications/{doc_id}/reject:
    post:
      tags: [知识库后台]
      operationId: rejectKbDocument
      summary: 驳回入库申请（意见必填）
      parameters:
        - { $ref: '#/components/parameters/DocIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [opinion]
              properties:
                opinion: { type: string, minLength: 5 }
      responses:
        '200':
          description: 已驳回
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbDocument' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 演示环境静态 token；生产替换为机构 SSO（替换成本见 04 §2）
  parameters:
    CaseIdPath:
      name: case_id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/CaseId' }
    DocIdPath:
      name: doc_id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 错误码（08 §6 错误码表）
          example: E-DISP-AUTH
        message: { type: string }
        trace_id: { type: string, description: OTel trace_id，联动可观测 }
    CaseId:
      type: string
      pattern: '^CASE-\d{8}-\d{4}$'
      example: CASE-20260813-0001
    CaseStatus:
      type: string
      description: 状态机（02 §7 + 08 §2 枚举字典，DA-INV-01 守护）
      enum: [REGISTERED, AGGREGATING, INVESTIGATING, PENDING_APPROVAL,
             APPROVED, REJECTED, MANUAL_REVIEW, DISPOSING, DISPOSED,
             VERIFIED, ROLLBACK, ARCHIVED]
...
    CaseSummary:
      type: object
      properties:
        case_id: { $ref: '#/components/schemas/CaseId' }
        subject_ref: { type: string }
        status: { $ref: '#/components/schemas/CaseStatus' }
        risk_score: { type: integer, minimum: 0, maximum: 100 }
        current_agent: { type: string, description: AA-AG-0x 编号 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        # 审批超时标记不在案件摘要上：BA-BR-13 升级写在 DA-T-07 escalated_at（API-W-08 返回）
    CaseDetail:
      allOf:
        - $ref: '#/components/schemas/CaseSummary'
        - type: object
          properties:
            context:
              type: object
              additionalProperties: true
              description: 共享状态（DA-T-03 context_json，AA-CL-03 载体）
            signal_count: { type: integer }
            evidence_count: { type: integer }
            matrix_room: { type: string, description: Matrix 房间链接 }
            trace_id: { type: string }
    RiskSignal:
      type: object
      description: DA-T-04 信号结构（四源统一，03 §4）
      properties:
        signal_id: { type: string }
        case_id: { $ref: '#/components/schemas/CaseId' }
        source: { type: string, enum: [tx, credit, sentiment, complaint] }
        type: { type: string, example: high_freq_small_amount }
        confidence: { type: number, minimum: 0, maximum: 1 }
        raw_ref: { type: string, description: 原始数据引用（tx_id 或外部源 id） }
        query_reason: { type: string, description: 查询事由（BA-BR-10 必填） }
        degraded: { type: boolean, description: 数据源降级标记 }
        velocity_json:
          type: object
          nullable: true
          description: 频次统计特征（BA-BR-14，对标社区 velocity 实践）；tx 源信号必填
          properties:
            velocity_1h: { type: object, description: "近 1 小时 {count, amount}" }
            velocity_24h: { type: object, description: "近 24 小时 {count, amount}" }
        ts: { type: string, format: date-time }
    VelocityFeature:
      type: object
      description: 频次统计特征（BA-BR-14，与 RiskSignal.velocity_json 同构）
      properties:
        velocity_1h:
          type: object
          properties:
            count: { type: integer, description: 近 1 小时交易笔数 }
            amount: { type: number, description: 近 1 小时交易金额 }
        velocity_24h:
          type: object
          properties:
            count: { type: integer, description: 近 24 小时交易笔数 }
            amount: { type: number, description: 近 24 小时交易金额 }
    AggregateResult:
      type: object
      description: AA-SK-01 聚合流水线结果（US-E3-03/04，分级裁决路由）
      properties:
        case_id: { $ref: '#/components/schemas/CaseId' }
        status: { $ref: '#/components/schemas/CaseStatus' }
        route:
          type: string
          enum: [noise, auto_release, investigate, all_fail]
          description: 裁决路由（降噪/自动放行/转调查/全源失败转人工）
        risk_score: { type: integer, minimum: 0, maximum: 100 }
        velocity: { $ref: '#/components/schemas/VelocityFeature' }
        signals_count: { type: integer }
        degraded_sources:
          type: array
          items: { type: string, enum: [tx, credit, sentiment, complaint] }
        exec_id: { type: string, nullable: true, description: 自动放行处置凭证（DA-T-06） }
    Evidence:
      type: object
      description: DA-T-05 证据条目（共享内核结构，01 §10.2）
      properties:
        evidence_id: { type: string }
        case_id: { $ref: '#/components/schemas/CaseId' }
        claim: { type: string, description: 结论主张 }
        source_ref: { type: string, description: 数据来源（含 RAG doc_id 引用） }
        confidence: { type: number, minimum: 0, maximum: 1 }
        ts: { type: string, format: date-time }
    Disposition:
      type: object
      description: DA-T-06 处置记录
      properties:
        exec_id: { type: string }
        case_id: { $ref: '#/components/schemas/CaseId' }
        action: { type: string, enum: [block, freeze, reduce, release] }
        amount: { type: number, format: double, description: 涉案/调整金额 }
        idempotency_key:
          type: string
          description: case_id+decision_type（AA-SK-03，DA-INV-03）
        approval_ref: { type: string, nullable: true }
        status:
          type: string
          enum: [submitted, executed, failed, rolled_back]
        ts: { type: string, format: date-time }
    Approval:
      type: object
      description: DA-T-07 审批工单（AG-04 创建、人类回填）
      properties:
        approval_id: { type: string }
        case_id: { $ref: '#/components/schemas/CaseId' }
        approver: { type: string, nullable: true }
        decision: { type: string, enum: [pending, approved, rejected] }
        opinion: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        decided_at: { type: string, format: date-time, nullable: true }
        requested_action:
          type: string
          enum: [block, freeze, reduce, release]
          nullable: true
          description: 建单请求的处置动作（C1 凭证验真比对项）
        requested_amount: { type: number, nullable: true, description: 涉案金额 }
        escalated_at:
          type: string
          format: date-time
          nullable: true
          description: BA-BR-13/SC-09 超时升级时刻（非空即前端标红，后台扫描每轮幂等写入）
    ReviewRequest:
      type: object
      required: [conclusion, opinion]
      properties:
        conclusion: { type: string, enum: [release, block, escalate] }
        opinion: { type: string, minLength: 5 }
    VerifyRequest:
      type: object
      required: [exec_id]
      properties:
        exec_id: { type: string, description: 待核验处置执行凭证（DA-T-06） }
    InvestigationResult:
      type: object
      description: AA-SK-02 调查结果（US-E4-01~03）
      properties:
        case_id: { $ref: '#/components/schemas/CaseId' }
        case_status: { $ref: '#/components/schemas/CaseStatus' }
        hypothesis:
          type: object
          properties:
            pattern: { type: string, description: 定性假设（跑分/盗卡/团伙盗刷/待定） }
            rule_basis: { type: string, description: 规则兜底依据 }
            citations:
              type: array
              items: { type: object, properties: { doc_id: { type: string }, title: { type: string }, similarity: { type: number } } }
            kb_note: { type: string, description: 未命中时显式声明"无库内匹配" }
        graph:
          type: object
          properties:
            nodes: { type: integer }
            edges: { type: integer }
            edge_types: { type: array, items: { type: string } }
            black_hits: { type: array, items: { type: string } }
        impact:
          type: object
          properties:
            accounts: { type: integer }
            amount_24h: { type: number }
        evidence_fixed: { type: boolean }
    VerificationResult:
      type: object
      description: AA-SK-04 核验结果（US-E6-01/02）
      properties:
        case_id: { $ref: '#/components/schemas/CaseId' }
        consistency_check: { type: boolean }
        trace_complete: { type: boolean }
        case_status: { $ref: '#/components/schemas/CaseStatus' }
        audit_report: { type: string, description: 核验一致时的审计报告（落 DA-T-05） }
        kb_application: { type: string, description: 复盘入库申请 doc_id（pending） }
        rollback_exec_id: { type: string, description: 核验不一致时的反向处置凭证 }
    DecideRequest:
      type: object
      required: [decision, opinion]
      properties:
        decision: { type: string, enum: [approve, reject] }
        opinion: { type: string, minLength: 5, description: 批准/驳回意见均必填（02 §7） }
...
    KbDocument:
      type: object
      description: DA-T-09 知识文档/入库申请
      properties:
        doc_id: { type: string }
        category: { type: string, enum: [case, regulation, runbook] }
        title: { type: string }
        content_summary: { type: string }
        status: { type: string, enum: [pending, published, rejected] }
        applicant: { type: string, description: 固定 AA-AG-05 }
        ts: { type: string, format: date-time }
    GraphNode:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [Account, Device, Transaction, PayeeAccount, IpRange, Contact, Case], description: DA-N-01~07 }
        risk_flag: { type: string, enum: [none, watch, gray, black] }
    GraphLink:
      type: object
      properties:
        source: { type: string }
        target: { type: string }
        relation: { type: string, enum: [SAME_DEVICE, SAME_PAYEE, SAME_IPSEG, SAME_CONTACT, INITIATED, INVOLVED_IN], description: DA-L-01~06 }
    GraphResponse:
      type: object
      properties:
        start: { type: string }
        hops: { type: integer }
        nodes: { type: array, items: { $ref: '#/components/schemas/GraphNode' } }
        links: { type: array, items: { $ref: '#/components/schemas/GraphLink' } }
...